refactor(auth): convert account to TS and class
Because:

* New routes and some account cleanup is needed for allowing a new
  account with no password. This requires multiple re-usable functions
  to be extracted from the existing create account handler.

This commit:

* Converts the account.js file to TypeScript and refactors the handlers
  into a class to make room for the new account creation route.

Ref #8735
