+++ packages/fxa-auth-server/package.json
@@ -118,6 +118,7 @@
     "@types/node-zendesk": "^2.0.2",
     "@types/request": "2.48.5",
     "@types/safe-regex": "1.1.2",
+    "@types/uuid": "^8.3.0",
     "@types/verror": "^1.10.4",
     "acorn": "^8.0.1",
     "audit-filter": "^0.5.0",
